Build 38
Updated Fog of War to have curved edges. Makes the Fog of War blend better with the rest of the game.

+++ core/src/com/jaredzhao/castleblitz/components/player/CameraComponent.java
@@ -5,8 +5,8 @@ import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 
 public class CameraComponent implements Component{
-    public int cameraHeight = 250;
-    public float scale = ((float) Gdx.graphics.getHeight() / (float)cameraHeight);
+    public float cameraHeight = 250;
+    public float scale = ((float) Gdx.graphics.getHeight() / cameraHeight);
     public float cameraWidth = Gdx.graphics.getWidth() / scale;
     public OrthographicCamera camera = new OrthographicCamera(cameraWidth, cameraHeight);
 }
